refactor(server): /api/version + claude-profile + update 라우트 분할 (Plan C #C9 2단계)
- src/server/routes/version.ts 신규: 4개 라우트 캡슐화 (version, claude-profile, refresh, update)
- getCurrentVersion 헬퍼를 라우트 모듈로 이동 (import.meta.url 기준 package.json)
- safeError 적용 (sanitizeErrorMessage(getErrorMessage(e)) → safeError)
- dashboard-api.ts에서 SelfUpdater/checkClaudeQuota/basename/fileURLToPath import 제거
- dashboard-api.ts 1977 → 1858줄 (-119)

남은 도메인: jobs/projects/config/setup/stats/repositories/health/doctor/sse
